+++ src/main/java/ua/duikt/learning/java/pro/spring/dtos/PatchAssigneeRequest.java
@@ -12,5 +12,5 @@ import lombok.Data;
 @Builder
 @AllArgsConstructor
 public class PatchAssigneeRequest {
-    private Integer assigneeId;
+    private Long assigneeId;
 }
