+++ src/main/java/ua/duikt/learning/java/pro/spring/dtos/PatchStatusRequest.java
@@ -12,5 +12,5 @@ import lombok.Data;
 @Builder
 @AllArgsConstructor
 public class PatchStatusRequest {
-    private Integer statusId;
+    private Long statusId;
 }
